Break up mdoc.ts
