+++ packages/server/src/formats/mdoc/index.ts
@@ -1,0 +1,1 @@
+export { verifyMdocPresentation } from './verifyMdocPresentation.ts';
