+++ packages/server/src/formats/mdoc/verifyMdocPresentation.ts
@@ -1,0 +1,51 @@
+import { decodeCBOR } from '@levischuck/tiny-cbor';
+
+import type { DecodedCredentialResponse } from './types.ts';
+
+/**
+ * Verify an mdoc presentation as returned through the DC API
+ */
+export async function verifyMdocPresentation(
+  responseBytes: Uint8Array,
+): Promise<VerifiedNamespace> {
+  const decodedResponse = decodeCBOR(responseBytes) as DecodedCredentialResponse;
+  const document = decodedResponse.get('documents')[0];
+
+  // TODO: The rest of the owl
+
+  return {};
+}
+
+/**
+ * A map of namespaces and their verified, issuer-signed element identifiers and values
+ *
+ * Example:
+ *
+ * ```
+ * {
+ *   "org.iso.18013.5.1": [
+ *     [ "given_name", "Jon" ],
+ *     [ "family_name", "Smith" ],
+ *     [ "age_over_21", true ]
+ *   ]
+ * }
+ * ```
+ */
+export type VerifiedNamespace = { [namespaceID: string]: [elemID: string, elemValue: unknown][] };
+
+const identifiers = [
+  'family_name',
+  'given_name',
+  'birth_date',
+  'issue_date',
+  'expiry_date',
+  'issuing_country',
+  'issuing_authority',
+  'document_number',
+  'portrait',
+  'un_distinguishing_sign',
+  'age_in_years',
+  'age_birth_year',
+  'age_over_NN',
+] as const;
+export type Identifier = typeof identifiers[number];
